Tidied up rename and rotate
