+++ pdf_bot/files/file.py
@@ -1,17 +1,14 @@
-import os
-import re
-import shutil
-import tempfile
-
-from telegram import ReplyKeyboardMarkup, ReplyKeyboardRemove
+from telegram import ReplyKeyboardMarkup
 from telegram.constants import MAX_FILESIZE_DOWNLOAD
 from telegram.ext import ConversationHandler, CommandHandler, MessageHandler, Filters
 from telegram.ext.dispatcher import run_async
 
 from pdf_bot.constants import WAIT_TASK, WAIT_DECRYPT_PW, WAIT_ENCRYPT_PW, WAIT_ROTATE_DEGREE, WAIT_SCALE_BY_X, \
     WAIT_SCALE_BY_Y, WAIT_SCALE_TO_X, WAIT_SCALE_TO_Y, WAIT_SPLIT_RANGE, WAIT_FILE_NAME, PDF_INFO
-from pdf_bot.utils import cancel, process_pdf, send_result_file
+from pdf_bot.utils import cancel
 from pdf_bot.files.crypto import ask_decrypt_pw, ask_encrypt_pw, decrypt_pdf, encrypt_pdf
+from pdf_bot.files.rename import ask_pdf_new_name, rename_pdf
+from pdf_bot.files.rotate import ask_rotate_degree, rotate_pdf
 from pdf_bot.files.scale import ask_scale_x, ask_scale_by_y, ask_scale_to_y, pdf_scale_by, pdf_scale_to
 from pdf_bot.files.split import ask_split_range, split_pdf
 from pdf_bot.photos import get_pdf_cover, get_pdf_photos, pdf_to_photos, process_photo
@@ -146,111 +143,5 @@ def receive_photo_task(update, context):
 
     if user_data[PHOTO_ID] == file_id:
         del user_data[PHOTO_ID]
-
-    return ConversationHandler.END
-
-
-@run_async
-def ask_pdf_new_name(update, _):
-    """
-    Ask and wait for the new file name
-    Args:
-        update: the update object
-        _: unused variable
-
-    Returns:
-        The variable indicating to wait for the file name
-    """
-    update.message.reply_text('Send me the file name that you\'ll like to rename your PDF file into.',
-                              reply_markup=ReplyKeyboardRemove())
-
-    return WAIT_FILE_NAME
-
-
-@run_async
-def rename_pdf(update, context):
-    """
-    Rename the PDF file with the given file name
-    Args:
-        update: the update object
-        context: the context object
-
-    Returns:
-        The variable indicating to wait for the file name or the conversation has ended
-    """
-    user_data = context.user_data
-    if PDF_INFO not in user_data:
-        return ConversationHandler.END
-
-    text = re.sub(r'\.pdf$', '', update.message.text)
-    invalid_chars = r'\/*?:\'<>|'
-    if set(text) & set(invalid_chars):
-        update.message.reply_text(f'File names can\'t contain any of the following characters:\n{invalid_chars}\n'
-                                  f'Send me another file name.')
-
-        return WAIT_FILE_NAME
-
-    new_fn = '{}.pdf'.format(text)
-    update.message.reply_text(f'Renaming your PDF file into *{new_fn}*', parse_mode='Markdown')
-
-    # Download PDF file
-    file_id, _ = user_data[PDF_INFO]
-    tf = tempfile.NamedTemporaryFile()
-    pdf_file = context.bot.get_file(file_id)
-    pdf_file.download(custom_path=tf.name)
-
-    with tempfile.TemporaryDirectory() as dir_name:
-        out_fn = os.path.join(dir_name, new_fn)
-        shutil.move(tf.name, out_fn)
-        send_result_file(update, out_fn)
-
-    # Clean up memory and files
-    if user_data[PDF_INFO] == file_id:
-        del user_data[PDF_INFO]
-    try:
-        tf.close()
-    except FileNotFoundError:
-        pass
-
-    return ConversationHandler.END
-
-
-@run_async
-def ask_rotate_degree(update, _):
-    """
-    Ask and wait for the rotation degree
-    Args:
-        update: the update object
-        _: unused variable
-
-    Returns:
-        The variable indicating to wait for the rotation degree
-    """
-    keyboard = [['90'], ['180'], ['270']]
-    reply_markup = ReplyKeyboardMarkup(keyboard, one_time_keyboard=True)
-    update.message.reply_text('Select the degrees that you\'ll like to rotate your PDF file in clockwise.',
-                              reply_markup=reply_markup)
-
-    return WAIT_ROTATE_DEGREE
-
-
-@run_async
-def rotate_pdf(update, context):
-    """
-    Rotate the PDF file with the given rotation degree
-    Args:
-        update: the update object
-        context: the context object
-
-    Returns:
-        The vairable indicating the conversation has ended
-    """
-    if PDF_INFO not in context.user_data:
-        return ConversationHandler.END
-
-    degree = int(update.message.text)
-    update.message.reply_text(f'Rotating your PDF file clockwise by {degree} degrees',
-                              reply_markup=ReplyKeyboardRemove())
-    process_pdf(update, context, 'rotated', rotate_degree=degree)
 
     return ConversationHandler.END
